Added first link, needs refactoring

+++ Models/MenuModel.js
@@ -16,7 +16,7 @@ Menu.init({
         primaryKey: true
     },
     price: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(4,2),
         allowNull: false,
     }
 }, {
